launch-client: Bind to a specific launcher-service instance
The launcher API is stateful: once we have asked an instance of the
launcher to start a subprocess, we must send any subsequent Terminate()
or SendSignal() messages to the original instance, not to a second
instance that has taken over the same well-known bus name with
`s-r-launcher-service --replace`. Similarly, we must listen for
ProcessExited signals from the original instance, not a replacement.

We can achieve this by resolving the well-known bus name (analogous to a
DNS name) to a unique name (analogous to a dynamic IP address, but more
so, since it is never reused during the lifetime of a dbus-daemon).

Signed-off-by: Simon McVittie <smcv@collabora.com>