launcher-service: Set $MAINPID in launched commands
If the launched process is run directly (as opposed to via a wrapper
script that runs it as a subprocess or puts it in the background),
then this is a convenient way to attach a debugging tool like strace
or gdbserver to the launched process (typically a game).

The `MAINPID` variable name is borrowed from systemd.exec(5).

Note that for tools that ptrace the game (like gdbserver and strace),
this will not work on Ubuntu or other distributions that use YAMA, unless
you set sysctl `kernel.yama.ptrace_scope=0` to remove the restriction
on debugging non-child processes.

main_pid remains set even after the main process has terminated,
because this gives us a way to keep track of whether the process we
are looking at was the main process (which we will want in a subsequent
commit to clarify the message when implementing `terminate_after`),
and to keep track of whether the main process was ever run (which
we will want for steamrt/tasks#129).

Resolves: steamrt/tasks#126
Signed-off-by: Simon McVittie <smcv@collabora.com>