launcher-service: Don't exit on name loss if children exist
Instead of using a GMainLoop, it's often clearer what is going on if we
explicitly loop until a condition becomes false, and that seems to be
the case here.

Keeping better track of the circumstances under which we exit revealed
that if we lose our D-Bus name (due to being replaced by a second
instance with --replace), we would have exited after a short delay.
This seems unfortunate, because we don't terminate our child processes
in that situation, but exiting from the launcher-service process means
that the launch-client processes cannot be informed when the child
processes exit or given the opportunity to send signals to them, even
if they are tracking the launcher-service by its unique name (as we
started to do in a previous commit). Resolve this by continuing to not
terminate the child processes, but also lingering until they have all
exited, so that we can propagate their exit statuses back to the caller
if it is watching our unique name.

Similarly, in signal_handler() and exit_on_readable_cb(), previously we
would unconditionally stop the main loop immediately. Now, we wait for
child processes to exit and for a possible queue of D-Bus calls to be
processed.

Related to steamrt/tasks#128

Signed-off-by: Simon McVittie <smcv@collabora.com>