launcher-service: Fall back to kill() if killpg() fails
If the process we are terminating is not a process group leader, then
killpg() will fail with ESRCH. This is particularly significant if we
are terminating a wrapped process, for which we set
`child_setup_data.keep_tty_session = TRUE` so that it will not give up
its controlling terminal (if any).

Before this change, `test_wrap_wait` in tests/pressure-vessel/launcher.py
would leak its `sleep 600` subprocess, which was not correctly
terminated by calling `s-r-launch-client --terminate`. Replace that
with a `cat` subprocess so that we can detect when it has been
terminated, and replace the previous use of stdout (which duplicated
similar code in `test_wrap_info_fd`) to facilitate that.

This will become more important when we keep better track of child
processes in order to avoid exiting while there are still managed child
processes running.

Signed-off-by: Simon McVittie <smcv@collabora.com>