launcher-service: Fail if bus names cannot be acquired
Previously, the listener object would emit ::ready (because all of its
zero remaining bus names have been acquired) and run the command as a
subprocess even though the listener object is not really in the intended
state. Now, if not run with --no-stop-on-name-loss, it will print an
error message the first time a name cannot be acquired and exit
unsuccessfully.

Helps: steamrt/tasks#129
Signed-off-by: Simon McVittie <smcv@collabora.com>