Add VA-API driver version information
For VA-API drivers we can expect to obtain the libva version they depend
on.

This is performed by looking for a symbol that starts with
"__vaDriverInit_x_", where x represents the VA-API major version.

The version check, when not needed, can be disabled with the flag
`SRT_CHECK_FLAGS_SKIP_SLOW_CHECKS`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>