pv-runtime: Clarify that we are intentionally hiding Vulkan layers
Mounting a tmpfs to "hide" vulkan/implicit_layer.d is necessary to stop
$XDG_DATA_HOME/vulkan/implicit_layer.d from being loaded in preference
to /overrides/share/vulkan/implicit_layer.d. However, this is not
necessarily obvious to pressure-vessel users attempting to work around
an issue like ValveSoftware/steam-runtime#511. Emit a log message that
makes it clearer that this is intentional.

Signed-off-by: Simon McVittie <smcv@collabora.com>