bin: Add a simpler interface to wrap a game in the launcher-service
If launcher-service is (mostly) mechanism, then this is policy.
It uses argv[1] and $STEAM_COMPAT_LAUNCHER_SERVICE to decide whether
to wrap a launcher-service around a game process. The idea is that
Steam will set STEAM_COMPAT_LAUNCHER_SERVICE to the most appropriate
layer of the stack at which to be inserting arbitrary debugging
commands into a game.

In previous prototypes, this was open-coded into each compat tool, but
centralizing it is a lot easier for compat tool authors. Put a version
in the name we use, to ensure that if we find we need to make breaking
changes, we can rename to a new interface version (and optionally keep
the old one around too).

To minimize its startup time cost, this doesn't use GLib, only glibc.

Signed-off-by: Simon McVittie <smcv@collabora.com>