docs: Clarify installation directory behaviour
When writing this document I wasn't sure whether the default working
directory was the top-level install folder or the directory containing
the executable, but by looking at Dota 2 we can see that the answer
is the top-level install folder: Dota 2's executable is in a
subdirectory, but its initial working directory is not.

Add some links to Steam partner documentation (previously only in the
slr-for-game-developers document) to make this a bit clearer.

Signed-off-by: Simon McVittie <smcv@collabora.com>