tests: Work around GLib < 2.70 not handling multi-line messages correctly
GLib older than 2.70 corrupts TAP output when a g_test_message()
contains newlines. For example, "hello\nworld" ends up as:

    # hello
    world

rather than the correct TAP diagnostic

    # hello
    # world

We are particularly susceptible to this when we output debug-log
messages, which sometimes contain commands' multi-line output.
We already send the output of the default log handler to stderr, but
the GLib test framework also turns log messages into TAP diagnostics,
for which we need to be prepared to split the lines ourselves.

Signed-off-by: Simon McVittie <smcv@collabora.com>