Stop supporting STEAM_COMPAT_LAUNCHER_SERVICE=1391110, etc.
The original design for injecting commands into container code was that
it was going to be based on Steam app-IDs, but that's awkward for the
compat tool authors (particularly Proton, which has a lot of branches
with their own app-IDs), so the plan has changed.

It is now meant to be based on the compatmanager_layer_name in
toolmanifest.vdf, which only needs to be unique within a stack that
makes sense to use together: for instance, SteamLinuxRuntime_anything
can share "container-runtime", because it only makes sense for any
given stack to include one of them. Similarly, all Proton versions
can share "proton". We already recognised those names as a developer
convenience, even when they were not going to be used in production,
so nothing more needs to be done to support them.

This partially reverts commit ee4b6c47da7f4b277ebde495de36200facea77bc
"Add app-ID-based triggering for steam-runtime-launcher-service".
The addition of the --steam-app-id option has not been reverted, for
backwards-compat, but it no longer does anything.

Signed-off-by: Simon McVittie <smcv@collabora.com>