launch-client: Default to --directory="", except for Flatpak
When we are behaving like flatpak-spawn(1), its behaviour of inheriting
the current working directory makes sense: flatpak-spawn is meant to
run a command with different sandboxing, but as though it was a child
of the current process. Also, the current working directory for the
flatpak-portal and flatpak-session-helper services is uninteresting
(99% of the time it will be the user's home directory), so it will not
usually be useful to inherit.

However, when we are inserting commands into a running game, it makes
more sense to inherit the game's current working directory, which is
non-trivial (usually but not always the game's top-level installation
directory). Change s-r-launch-client so it does that by default,
instead of forcing its own working directory.

steamrt/tasks#145

Signed-off-by: Simon McVittie <smcv@collabora.com>