launch-client: Update release status
As of 2022-08-25, the copy of scout in the GA Steam Client has a
suitable version of s-r-launch-client, and the default branches of
soldier, sniper, scout-on-soldier, Proton 7.0 and Proton experimental
include the integration glue for s-r-launcher-interface-0. This lets us
remove various warnings and caveats.

Proton 6.0 and 5.13 will presumably never get this.

Signed-off-by: Simon McVittie <smcv@collabora.com>