tests: Write out JSON files so we can diff them nicely
Comparing two giant strings of JSON in g_assert_cmpstr() output is not
straightforward, so let's try to write them out to files and run diff
on them.

The AUTOPKGTEST_ARTIFACTS variable set by Debian's autopkgtest
framework is as good a mechanism as any for signalling where would be a
good place; or if that's not set, use a temporary directory.

Signed-off-by: Simon McVittie <smcv@collabora.com>