graphics-drivers: Add ability to search for drivers in hard-coded paths
At the moment the steam-runtime-tools helpers are built only for the
i386 and x86_64 architectures.
This means that we can only search for graphics libraries for these two
architectures.

With the addition of the flag "SRT_CHECK_FLAGS_NO_HELPERS", we hard-code
some of the most common paths where we expect to find the graphics
libraries and we skip using the helpers binaries.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>