Add support for FEX-Emu graphics thunking
When we are using an interpreter like FEX, we have a FEX's rootfs, that
is usually x86-64, and the real host system root, that is usually
AArch64.

When setting up the runtime we can also include the real host graphics
DRIs/layers to support thunking. Then, based on what the loader needs,
the correct architecture libraries will be loaded at runtime.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>