tests: Remove problematic check in mock executable
Returning an error when a library was located in a directory that ended
with 32, for x86_64, and 64, for i386, was, in theory, a good and simple
strategy.

However we also check libraries by using their `/proc/${PID}/fd/...`
path instead of the canonical absolute path.
This means that if the PID ended with 32 or 64, we could have received
an unexpected reply.

Given that this check is not actually necessary for the tests that we
have, we can simply remove it.

Fixes: https://gitlab.steamos.cloud/steamrt/tasks/-/issues/162

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>