pv-runtime: Don't capture libudev.so.0 for now
Capturing libudev.so.0 is desirable on systems that are using the
libudev0-shim originating from Arch Linux, which is an optional package
in at least Arch and Debian that implements the old libudev.so.0 ABI in
terms of the newer libudev.so.1 ABI. On these systems, it will take
precedence over the one from Steam Runtime 1 'scout' and will be more
likely to work in practice than scout's copy.

However, doing this can cause problems on systems where a user has made
libudev.so.0 a symlink to libudev.so.1. That's really misconfiguration,
because they're not compatible, but some people do this anyway and in
practice it often works.

While we work this out, partially revert commit e7175a68
"pv-runtime: Consider using provider libudev even if no dependency"
so that we will only use libudev.so.1 from the host system, and not
libudev.so.0.

Helps: steam-runtime-tools#96
Helps: https://github.com/ValveSoftware/steam-runtime/issues/533
Signed-off-by: Simon McVittie <smcv@collabora.com>