capture-libs: Add exact-soname option
Add a stricter version of `soname:`, where a library is considered to be
a match only if its DT_SONAME is exactly what capture-libs was asked to
look into.

This can be used as a protection against misconfigured systems where two
incompatible libraries have been symlinked.

E.g. if there is an unexpected symlink `libudev.so.0 -> libudev.so.1`,
with `soname:libudev.so.0` we will import `libudev.so.0` inside the
container even if its DT_SONAME is `libudev.so.1`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>