pv-runtime: Optionally log /overrides or full container contents
Logging the contents of `/overrides` enlarges the log by around 10% (on
a Debian 12 system) which we could consider doing by default, but for now
I've made it an opt-in.

Logging the entire contents of the container to be used is more like
a 300% increase, so it certainly needs to be an opt-in that we ask for
if it is genuinely going to be useful.

Helps: steamrt/tasks#187, steam-runtime-tools#97
Signed-off-by: Simon McVittie <smcv@collabora.com>