pv-wrap: Use correct bwrap exports while using an interpreter
When running under FEX-Emu on aarch64, because pv-wrap is an x86
process, the Flatpak code that deals with X11, D-Bus etc. implicitly
searches in both the FEX-Emu rootfs and the real root filesystem (in
that order) due to FEX-Emu's overlayfs-like behaviour.

But because bwrap's mount(2) doesn't do an automatic remap of the paths,
a workaround was manually implemented. Every time a file/directory was
available in the interpreter root, we adjusted the bind to use the
FEX-Emu's rootfs.

However this was not enough because, in the case where a path existed in
both the real host and the interpreter, we would have mounted only the
version from the interpreter.

In order to correctly handle all use cases we do the following:
- Path that only exists in the real host: keep it as is
- Path that only exists in the FEX rootfs: change the destination to
prefix the $PV_RUNTIME_PATH_INTERPRETER_ROOT. This will avoid exposing a
path that would not have been existed outside the container
- Path that is both in the FEX rootfs and the real host: duplicate it,
keeping one copy as is and editing the other to point the source to the
FEX rootfs and the destination to the $PV_RUNTIME_PATH_INTERPRETER_ROOT

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>