pv-exports: Treat /run/gfx the same as /run/host
If the /overrides hierarchy contains a symlink to /run/gfx/something,
we expect it to point into the /run/gfx that we already plan to create,
so there is no need to tell FlatpakExports to look for /run/gfx/something
on the host system. We already had similar logic for /run/host.

Signed-off-by: Simon McVittie <smcv@collabora.com>