pv-runtime: Abbreviate the unnecessary files we remove in debug log
We don't need to give the full path to the temporary mutable sysroot:
it's enough to say something like

    Deleting tmp-*/usr/lib/i386-linux-gnu/libc.so.6 because ...

As well as making the log smaller, with shorter lines, this will remain
consistent between similar runs, making it easier to see differences.

We continue to use the full path in errors and warnings, on the basis
that errors and warnings don't usually happen, and could have been caused
by a problem with any path component (whether before or after the part
that we exclude from debug messages).

Signed-off-by: Simon McVittie <smcv@collabora.com>