pv-runtime: Abbreviate path to generated JSON manifests
If we give a relative path starting with `overrides/`, it's reasonably
obvious from context what file we're writing to on the host system and
where it will end up in the container. The necessary code restructuring
to do this also makes it more obvious that the path `write_to_file` in
the current execution environment, the path `json_in_container` in the
resulting container, and the debug-log message all match up.

We still need `write_to_dir` in `setup_each_json_manifest()`, but only
to create the directory we're going to be populating.

Signed-off-by: Simon McVittie <smcv@collabora.com>