pv-runtime: Capture libraries into a path relative to /overrides
The resulting /proc-relative path will be a lot shorter than a full path
including /overrides, and if we are doing the same operations in the
same order during initialization, then in practice the file descriptor
will be the same; both of these make log files easier to diff.

This also lets us thread relative paths through the various functions
that call this one, which will make it easier to make their logging
equally concise.

Signed-off-by: Simon McVittie <smcv@collabora.com>