pv-runtime: Open the overrides directory as a file descriptor
We'll use this to do I/O relative to this file descriptor, instead of
using absolute paths in the current filesystem namespace.

To make this work, we have to create the directory with its parents.
Previously, we were only creating the last path component, which was
fine when creating /tmp/pressure-vessel-wrap.XXXXXX/overrides in a
directory /tmp/pressure-vessel-wrap.XXXXXX that we already created, but
didn't work in practice when creating
.../var/tmp-XXXXXX/usr/lib/pressure-vessel/overrides. This was hidden by
the fact that we didn't check the result of g_mkdir() for errors, and by
the time we tried to do anything with the overrides directory, we had
created .../var/tmp-XXXXXX/usr/lib/pressure-vessel/overrides/lib/TUPLE
with its parents (with the side-effect of creating the overrides
directory).

Signed-off-by: Simon McVittie <smcv@collabora.com>