pv-wrap: Fix /run/host binding when under an emulator
When using an emulator we will end up with a container that has
`/run/host`, which points to the real host root, and
`/run/pressure-vessel/interpreter-root/run/host`, that is instead the
interpreter rootfs.

However we can't populate `/run/host` with `flatpak_exports_add_host*`
and rely on `append_adjusted_exports()` to correctly handle the
adjusting for the interpreter mount points. Some information will be
inevitably lost, because the real host and the interpreter rootfs don't
necessarily have the same structure.

Instead, to avoid this information loss, we explicitly mount `/etc` and
use `pv_bwrap_bind_usr()`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>