build: Stop including source code in pressure-vessel-relocatable.deb
This makes the .deb considerably smaller, around 350K rather than 1.3M.
If we want the source code, we can build the .dsc like any other dpkg
package.

The copy of the source code in /usr used to be useful when we constructed
a relocatable pressure-vessel installation by installing it into the
Steam Runtime 'scout' SDK and then repacking it from there, but that was
rather slow, so our release pipeline is now a lot more direct than that.

Signed-off-by: Simon McVittie <smcv@collabora.com>