populate-depot: Use .tarball-version to determine scripts version
This allows populate-depot.py to learn its own version number when used
by unpacking a steam-runtime-tools source tarball, rather than from git,
and write that into VERSIONS.txt as the "scripts" component.

Now that we've combined SteamLinuxRuntime and s-r-t into one repository,
the "scripts" component is somewhat redundant for the soldier and
sniper runtimes (since it will always match up with pressure-vessel),
but it's still interesting for the scout-on-soldier runtime (which does
not contain its own copy of pressure-vessel).

Signed-off-by: Simon McVittie <smcv@collabora.com>