scout-on-soldier: Use /bin/bash for the main script
Because this script runs inside the SteamLinuxRuntime_soldier container,
we don't need to worry about whether /bin/bash exists or makes sense:
we know that it should be the bash from Debian 10, which is what this
script was written to target. Use it by its absolute path instead of
doing a PATH search, to add a bit more predictability.

Signed-off-by: Simon McVittie <smcv@collabora.com>