inspect-library: Parse ELF header for version definitions
If in the `.symbols` file there is a verdef `$VERSION@$VERSION`,
we can manually parse the ELF header to check if the library has
the expected `$VERSION` or not.

All missing versions will be reported in output in a list called
`missing_versions`.

If the library is unexpectedly unversioned, the
`unexpectedly_unversioned` entry will be set to `true`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>