runtime: Fix path_visible_in_provider_namespace() check
The check to ensure that the path had the prefix `sbin` was wrongly
using the OR comparison instead of the AND.

This meant that `path_visible_in_provider_namespace()` was returning
true for every path that was under a 4 characters long directory. E.g.
`home/[...]` was unexpectedly considered to be visible in the provider
namespace.

This check is used in `pv_runtime_take_from_provider()` to evaluate if a
simple symlink would be enough or if instead the files should be copied
in a directory that is visible from inside the container. Leading to
possible unexpected dangling symlinks.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>