CI: Build and upload content for Steam app 1070560 (SteamLinuxRuntime)
Ever since we started using the "layered" scout-on-soldier runtime
(depot version 0.20210630.32, Steampipe build ID 6958552), the content
of SteamLinuxRuntime has depended only on files shipped by
steamrt/steamlinuxruntime>, which was merged into this project in
v0.20221208.0.

This means it's no longer useful to build SteamLinuxRuntime as part of
the same workflows as the soldier and sniper runtimes, so we can
simplify its workflow by having its version advance in lockstep with
pressure-vessel.

Helps: steamrt/tasks#200, steamrt/tasks#207
Signed-off-by: Simon McVittie <smcv@collabora.com>