runtime: Move mkdir into setup_json_manifest()
Move the creation of the directory in the same function that actually
uses it. This will make it more clear what is happening and reduces the
function implicit requirements.

This is also a preparation step towards reusing the manifest basename,
that will be added as a subsequent commit.

The directories are only created when we have an ICD_KIND_ABSOLUTE
because, in all the other cases, the eventual need to create the parent
directories is automatically handled by
`pv_runtime_take_from_provider()`.

With this commit we might end up with a slightly increased number of
`g_mkdir_with_parents()` calls, but the overhead should be negligible,
compared to the complexity of pressure-vessel.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>