pressure-vessel: Reuse the drivers basename
When creating the JSON manifests, we were simply naming them
sequentially, as an easy way to avoid conflicts.

However, since Vulkan Loader 1.3.234, it is possible to select which
driver to use by listing their names in the environment variable
`VK_LOADER_DRIVERS_SELECT`. This functionality would not work in
pressure-vessel, unless we keep the original JSON manifest name.

With this commit now we always reuse the JSON manifest basename.
In the rare events where this could cause a conflict, we create
unique sub directories to avoid issues.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>