inspect-library: Remove the older JSON based output
In `inspect-library` we only care about the line based output.

With this commit we make the line based format the default and remove
the older JSON based one. The option `--line-based` is still available
in `inspect-library` to avoid any unexpected breakage and we expect to
remove it after a couple of Steam Runtime releases.
In `inspect-library-libelf` instead, we already removed the
`--line-based` option because it is a new helper and there are no
backwards-compat concerns.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>