library: Separate the libelf based checks into a separate helper
`libelf.so.1` is usually always available in the host system because it
is a dependency of several core packages, e.g. mesa.

However on some less traditional distributions, like NixOS or Gentoo,
this might not always be the case.

We can't compile `inspect-library` with an `RPATH` because we want to
load libraries as the host system does. And this means that
`libelf.so.1` needs to be picked from the host system and we can't use
the version included in pressure-vessel.

For this reason we move the libelf based checks into a separate helper
and keep that dependency out of the critical path to identify the
graphics drivers.

Effectively, this commit brings back `inspect-library.c` to the version
we had with af6b41e0d4edd0279e6dfe7920592a87f3d561a9.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>