CI: Only explicitly test populate-depot.py on Debian
We have two use-cases for this script: we use it to build production
runtime releases, which we can do in an environment that we control,
and we use it from tests/pressure-vessel/prepare-test-depots.sh as
part of testing pressure-vessel.

For the first use-case, we don't need to care about Ubuntu or Arch:
we can use any OS image of our choice, and in this case, we have
chosen to use Debian.

For the second use-case, we don't need to test the script exhaustively:
as long as tests/pressure-vessel/prepare-test-depots.sh succeeds, we're
happy with that.

Helps: steamrt/tasks#248
Signed-off-by: Simon McVittie <smcv@collabora.com>